fix(e2e): split e2e_test_shared.dart to satisfy 1000-line gate (Refs #2336)
Extract bootstrap/PNG helpers (e2eBootstrapNewGameToMap, e2eDecodePngAssetPathsParallel,
e2eEnsureAllRelocated64pxPngsLoad*, e2eCollectTextPreorder, etc.) into
e2e_test_shared_bootstrap.dart (237 non-comment lines). The main shared helper
file drops from 1085 to 854 non-comment lines, satisfying repo.dart_file_non_comment_line_size.
e2e_helpers.dart barrel imports both parts; no public API changes.
